requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
openai==0.28.0